requirements.txt
python-docx
PyPDF2
openai
python-dotenv